(WIP) Write support revisions & bugfixes

+++ webClient/src/app/shared/editor-control/editor-control.service.ts
@@ -216,7 +216,7 @@ export class EditorControlService implements ZLUX.IEditor, ZLUX.IEditorMultiBuff
 
   //almost like selectfilehandler, except altering the list of opened files
   public openFileHandler(fileContext: ProjectContext) {
-    console.log(`openFileHandler File ${fileContext.name}\n`);
+    this.log.debug(`openFileHandler File ${fileContext.name}\n`);
     for (const file of this._openFileList.getValue()) {
       file.opened = false;
       file.active = false;
@@ -234,7 +234,7 @@ export class EditorControlService implements ZLUX.IEditor, ZLUX.IEditorMultiBuff
   }
 
   public closeFileHandler(fileContext: ProjectContext) {
-    console.log(`closeFileHandler start\n`);
+    this.log.debug(`closeFileHandler start\n`);
     let cacheFileName = `${fileContext.model.fileName}:${fileContext.model.path}`;
     this.previousSessionData.stateCache = stateCache;
     this.previousSessionData._openFileList = this._openFileList.getValue();
@@ -243,29 +243,29 @@ export class EditorControlService implements ZLUX.IEditor, ZLUX.IEditorMultiBuff
       delete stateCache[cacheFileName];
     }
 
-    /* dequeue dataset */
-    console.log(`closeFileHandler model File name is ${fileContext.model.fileName}`);
-    
-          /* Send DEQ dataset  */
-          let filePath = ['/', '\\'].indexOf(fileContext.model.path.substring(0, 1)) > -1 ? fileContext.model.path.substring(1) : fileContext.model.path;
-          const enqRequestUrl = ZoweZLUX.uriBroker.datasetEnqueueUri(filePath);   /* the ENQ URL */
-          let _observable = this.http.delete(enqRequestUrl);                        /* prepare to delete the ENQ */
+    if (fileContext.model.isDataset) {
+      /* dequeue dataset */
+      this.log.debug(`closeFileHandler model File name is ${fileContext.model.fileName}`);
+      
+      /* Send DEQ dataset  */
+      let filePath = ['/', '\\'].indexOf(fileContext.model.path.substring(0, 1)) > -1 ? fileContext.model.path.substring(1) : fileContext.model.path;
+      const enqRequestUrl = ZoweZLUX.uriBroker.datasetEnqueueUri(filePath);   /* the ENQ URL */
+      let _observable = this.http.delete(enqRequestUrl);                        /* prepare to delete the ENQ */
 
-          console.log(`closeFileHandler delete request is for ${fileContext.model.fileName}`);
+      this.log.debug(`closeFileHandler delete request is for ${fileContext.model.fileName}`);
 
-          /* subscribe to the DELETE request */  
-          _observable.subscribe({
-            next: (response: any) => {
-              console.log(`closeFileHandler delete request OK`);              
-            },
-            error: (err) => {
-              // console.log(`closeFileHandler delete request FAILED, error status ${err.status}`); 
-              console.log(`closeFileHandler delete request FAILED.  Error is: `, err);             
-            }
-          });
-          /* end of subscribe */
-          
-    /* end of dequeue dataset */
+      /* subscribe to the DELETE request */  
+      _observable.subscribe({
+        next: (response: any) => {
+          this.log.debug(`closeFileHandler delete request OK`);              
+        },
+        error: (err) => {
+          // console.log(`closeFileHandler delete request FAILED, error status ${err.status}`); 
+          this.log.warn(`closeFileHandler delete request FAILED.  Error is: `, err);             
+        }
+      });
+    }
+
     !fileContext.opened ? this.log.warn(`File ${fileContext.model.fileName} already closed.`) : fileContext.opened = false;
     !fileContext.active ? this.log.warn(`File ${fileContext.model.fileName} already inactive.`) : fileContext.active = false;
     fileContext.changed = false;
